Cleanup and flip math operation order
Math operation order on the stack should follow Reverse Polish Notation
where the order is: first_elt_pushed op second_elt_pushed. This was the
opposite of the element order used. This cuts down on the amount of
swapping needed in order to properly execute order-sensitive operations
such as subtract and divide in most instances.